+++ source/extensions/omni.isaac.core_nodes/python/impl/__init__.py
@@ -6,3 +6,4 @@ The Python bindings are all imported so that they can be used in the omni.graph.
 # One line per import is used to make them easier to read and find, grouped by originating file
 from .extension import Extension
 from .base_reset_node import BaseResetNode
+from .base_writer_node import BaseWriterNode, WriterRequest